backend/requirements.txt
asgiref==3.7.1
dj-database-url==2.0.0
Django==4.2.1
djangorestframework==3.14.0
psycopg2-binary==2.9.6
python-decouple==3.8
pytz==2023.3
requests==2.31.0
mysqlclient==2.2.0
mysql-connector-python==8.0.26
